Put Middleware in a seperate js file to avoid confusion
