+++ server/api/gateKeepingMiddleware.js
@@ -1,0 +1,22 @@
+const express = require('express');
+const router = express.Router();
+const {
+  models: { User },
+} = require('../db');
+
+// MIDDLEWARE
+router.use(express.json());
+
+// MIDDLEWARE FUNC
+const requireToken = async (req, res, next) => {
+  try {
+    const token = req.headers.authorization;
+    const user = await User.findByToken(token);
+    req.user = user;
+    next();
+  } catch (error) {
+    next(error);
+  }
+};
+
+module.exports = { requireToken };
